Android plugin/bridge - change API’s to assurance
